Add controlIdExists method to CompanySercice class and use this method in several other methods of this class

+++ src/main/java/org/becode/projects/services/CompanyService.java
@@ -24,21 +24,37 @@ public class CompanyService {
 	}
 	
 	public Company getSpecificCompany(int id) {
-		return rep.getById(id);
+		if(controlIdExists(id)) {
+			return rep.getById(id);			
+		}
+		throw new IllegalArgumentException(String.format("Company with id %d doesn't exist", id));
 	}
 	
 	public String createNewCompany(Company company) {
+		if(controlIdExists(company.getId())) {
+			return "Company with this id already exists";
+		}
 		rep.save(company);
 		return "New company successfully created";
 	}
 	
 	public String deleteCompany(int id) {
-		rep.deleteById(id);
-		return String.format("Company with id %d successfully removed", id);
+		if(controlIdExists(id)) {
+			rep.deleteById(id);
+			return String.format("Company with id %d successfully removed", id);			
+		}
+		return "Company with this id doesn't exists";
 	}
 	
 	public String updateCompany(Company company) {
-		rep.save(company);
-		return "Company successfully updated";
+		if(controlIdExists(company.getId())) {
+			rep.save(company);
+			return "Company successfully updated";			
+		}
+		return "Company with this id doesn't exists";
+	}
+	
+	private boolean controlIdExists(int id) {
+		return rep.existsById(id);
 	}
 }
